+++ app/Http/Controllers/Dashboard/NotificationController.php
@@ -8,12 +8,15 @@ use App\Http\Requests\UpdateNotificationRequest;
 use App\Models\Bank;
 use App\Models\SystemNotification;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Mcamara\LaravelLocalization\Facades\LaravelLocalization;
 
 class NotificationController extends Controller
 {
     public function index()
     {
+        Gate::authorize('list-notification');
+
         $notifications = SystemNotification::commonFilters([
             'search' => ['translation.title'],
             'status' => 'status',
@@ -24,28 +27,38 @@ class NotificationController extends Controller
 
     public function create()
     {
+        Gate::authorize('create-notification');
+
         $bankOptions = Bank::whereActivated()->translatedPluck('title', 'id');
         return view('dashboard.notifications.create', compact('bankOptions'));
     }
 
     public function store(StoreNotificationRequest $request)
     {
+        Gate::authorize('create-notification');
 
         $notification = SystemNotification::create($request->validated());
 
         return redirect()->route('dashboard.notifications.index')->with('success', __('dashboard.messages.success.created', ['resource' => $notification->title]));
     }
 
-    public function show(SystemNotification $notification) {}
+    public function show(SystemNotification $notification)
+    {
+        Gate::authorize('show-notification');
+    }
 
     public function edit(SystemNotification $notification)
     {
+        Gate::authorize('update-notification');
+
         $bankOptions = Bank::whereActivated()->translatedPluck('title', 'id');
         return view('dashboard.notifications.edit', compact('notification', 'bankOptions'));
     }
 
     public function update(UpdateNotificationRequest $request, SystemNotification $notification)
     {
+        Gate::authorize('update-notification');
+
         $data = $request->validated();
 
         foreach (LaravelLocalization::getSupportedLocales() as $localeCode => $properties) {
@@ -73,28 +86,9 @@ class NotificationController extends Controller
 
     public function destroy(SystemNotification $notification)
     {
+        Gate::authorize('delete-notification');
+
         $notification->delete();
         return redirect()->route('dashboard.notifications.index')->with('success', __('dashboard.messages.success.deleted', ['resource' => $notification->title]));
-    }
-
-    public function getBanknotifications(Request $request)
-    {
-        $bankId = $request->get('id');
-
-        if (!$bankId) {
-            return response()->json([]);
-        }
-
-        $notifications = SystemNotification::where('bank_id', $bankId)
-            ->whereActivated()
-            ->translatedPluck('title', 'id')
-            ->toArray();
-
-        $results = [];
-        foreach ($notifications as $id => $title) {
-            $results[] = ['id' => $id, 'text' => $title];
-        }
-
-        return response()->json($results);
     }
 }
