+++ app/Http/Controllers/Dashboard/RoleController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\UpdateRoleRequest;
 use App\Models\Permission;
 use App\Models\Role;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Validation\Rule;
 
 class RoleController extends Controller
@@ -17,6 +18,8 @@ class RoleController extends Controller
      */
     public function index()
     {
+        Gate::authorize('list-role');
+
         $roles = Role::commonFilters([
             'search' => ['translation.title'],
         ])->commonPaginate();
@@ -29,7 +32,9 @@ class RoleController extends Controller
      */
     public function create()
     {
-        $permissionOptions = Permission::translatedPluck('title');
+        Gate::authorize('create-role');
+
+        $permissionOptions = Permission::all()->groupBy('group');
         return view('dashboard.roles.create', compact('permissionOptions'));
     }
 
@@ -38,7 +43,10 @@ class RoleController extends Controller
      */
     public function store(StoreRoleRequest $request)
     {
+        Gate::authorize('create-role');
+
         $role = Role::create($request->validated());
+        $role->permissions()->sync($request->permission_ids);
 
         return redirect()->route('dashboard.roles.index')->with('success', __('dashboard.messages.success.created', ['resource' => $role->title]));
     }
@@ -48,6 +56,7 @@ class RoleController extends Controller
      */
     public function show(Role $role)
     {
+        Gate::authorize('show-role');
         // return view('dashboard.roles.show');
     }
 
@@ -56,7 +65,9 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
-        $permissionOptions = Permission::translatedPluck('title');
+        Gate::authorize('update-role');
+
+        $permissionOptions = Permission::all()->groupBy('group');
 
         return view('dashboard.roles.edit', compact('role', 'permissionOptions'));
     }
@@ -66,7 +77,10 @@ class RoleController extends Controller
      */
     public function update(UpdateRoleRequest $request, Role $role)
     {
+        Gate::authorize('update-role');
+
         $role->update($request->validated());
+        $role->permissions()->sync($request->permission_ids);
 
         return redirect()->route('dashboard.roles.index')->with('success', __('dashboard.messages.success.updated', ['resource' => $role->title]));
     }
@@ -76,6 +90,8 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        Gate::authorize('delete-role');
+
         $role->delete();
         return redirect()->route('dashboard.roles.index')->with('success', __('dashboard.messages.success.deleted', ['resource' => $role->title]));
     }
